+++ software/js/packages/example-webserial-timeline/package.json
@@ -14,12 +14,12 @@
         "@types/node": "^16.18.4",
         "@types/react": "^18.0.25",
         "@types/react-dom": "^18.0.9",
-        "crc-32": "^1.2.0",
         "lodash": "^4.17.21",
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
         "react-scripts": "^5.0.1",
         "smartknobjs-proto": "^1.0.0",
+        "smartknobjs-webserial": "^1.0.0",
         "typescript": "^4.9.3",
         "web-vitals": "^2.1.4"
     },
