+++ software/js/packages/smartknobjs-core/.eslintrc
