+++ software/js/packages/smartknobjs-core/.prettierrc
