+++ software/js/packages/smartknobjs-webserial/src/index.ts
@@ -1,0 +1,65 @@
+import {ProtoDecoderStream} from './streams/proto-decoder'
+import {DelimiterChunkedStream} from './streams/delimiter-transform'
+
+import {MessageCallback, SmartKnobCore} from 'smartknobjs-core'
+
+export class SmartKnobWebSerial extends SmartKnobCore {
+    private port: SerialPort | null
+    private writer: WritableStreamDefaultWriter<Uint8Array> | undefined = undefined
+
+    constructor(port: SerialPort, onMessage: MessageCallback) {
+        super(onMessage, (packet: Uint8Array) => {
+            this.writer?.write(packet).catch((e) => {
+                console.error('Error writing serial', e)
+                this.port?.close()
+                this.port = null
+                this.portAvailable = false
+            })
+        })
+        this.port = port
+        this.portAvailable = true
+        this.port.addEventListener('disconnect', () => {
+            console.log('shutting down on disconnect')
+            this.port = null
+            this.portAvailable = false
+        })
+    }
+
+    public async openAndLoop() {
+        if (this.port === null) {
+            return
+        }
+        await this.port.open({baudRate: SmartKnobCore.BAUD})
+        if (this.port.readable === null || this.port.writable === null) {
+            throw new Error('Port missing readable or writable!')
+        }
+
+        // TODO: even though stream transformers are clean, this should probably be converted back to
+        // raw byte-array processing within smartknobjs-core to avoid leaking cobs/crc/pb decoding
+        // from the core abstraction
+        const pbDecoder = new ProtoDecoderStream()
+        this.port.readable.pipeThrough(new DelimiterChunkedStream(0)).pipeTo(pbDecoder.writable)
+        const reader = pbDecoder.readable.getReader()
+        try {
+            this.writer = this.port.writable.getWriter()
+            try {
+                // eslint-disable-next-line no-constant-condition
+                while (true) {
+                    const {value, done} = await reader.read()
+                    if (done) {
+                        break
+                    }
+                    if (value !== undefined) {
+                        this.handleMessage(value)
+                    }
+                }
+            } finally {
+                console.log('Releasing writer')
+                this.writer?.releaseLock()
+            }
+        } finally {
+            console.log('Releasing reader')
+            reader.releaseLock()
+        }
+    }
+}
