--- software/js/packages/smartknobjs/src/typings/cobs.d.ts
@@ -1,4 +1,0 @@
-declare module 'cobs' {
-    export function decode(buf: Buffer): Buffer
-    export function encode(buf: Buffer, zeroFrame?: boolean): Buffer
-}
